requirements.txt
email-validator==2.1.0.post1
fastapi==0.108.0
motor==3.3.2
pydantic==2.5.3
pymongo==4.6.1
python-dotenv==1.0.0
uvicorn==0.25.0
